fixed progress indication during merge-road-nodes with >32k nodes
git-svn-id: http://vtp.googlecode.com/svn/trunk@4392 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/RoadFix.cpp
@@ -74,7 +74,7 @@ int RoadMapEdit::MergeRedundantNodes(bool bDegrees, bool progress_callback(int))
 	int removed = 0;
 
 	int nodes = NumNodes();
-	float total = nodes * nodes / 2;
+	float total = (float)nodes * nodes / 2;
 	int count1 = 0, count = 0, count_tick, count_last_tick = 0;
 	float tick_size = (float)total / 100;
 	double tolerance, tolerance_squared;
